Remove ApplicantService interface in favor of a concrete class.
